Services modified to get variations from YAML or env. var. Idem, DC geometry and start time flag.
