+++ reconstruction/dc/src/main/java/org/jlab/service/dc/DCHBEngine.java
@@ -1,19 +1,11 @@
 package org.jlab.service.dc;
 
-import cnuphys.magfield.MagneticFields;
 import cnuphys.snr.NoiseReductionParameters;
 import cnuphys.snr.clas12.Clas12NoiseAnalysis;
 import cnuphys.snr.clas12.Clas12NoiseResult;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
-import org.jlab.clas.reco.ReconstructionEngine;
-import org.jlab.detector.base.DetectorType;
-import org.jlab.detector.base.GeometryFactory;
-import org.jlab.detector.geant4.v2.DCGeant4Factory;
-import org.jlab.geom.base.ConstantProvider;
 import org.jlab.io.base.DataBank;
 import org.jlab.io.base.DataEvent;
 import org.jlab.io.hipo.HipoDataSource;
@@ -40,111 +32,28 @@ import org.jlab.rec.dc.track.TrackCandListFinder;
 import org.jlab.rec.dc.trajectory.DCSwimmer;
 import org.jlab.rec.dc.trajectory.RoadFinder;
 import org.jlab.rec.dc.trajectory.Road;
-import org.jlab.utils.CLASResources;
 import org.jlab.utils.groups.IndexedTable;
 
-public class DCHBEngine extends ReconstructionEngine {
+public class DCHBEngine extends DCEngine {
 
     String FieldsConfig="";
     AtomicInteger Run = new AtomicInteger(0);
-    DCGeant4Factory dcDetector;
-    String clasDictionaryPath ;
-    
+    //DCGeant4Factory dcDetector;
+    //String clasDictionaryPath ;
+
     public DCHBEngine() {
-        super("DCHB","ziegler","4.0");
-    }
-    
-    /**
-     * 
-     * determine torus and solenoid map name from yaml, else env, else crash
-     */
-    private void initializeMagneticFields() {
-        String torusMap=this.getEngineConfigString("torusMap");
-        String solenoidMap=this.getEngineConfigString("solenoidMap");
-        if (torusMap!=null) {
-            System.out.println("["+this.getName()+"] Torus Map chosen based on yaml: "+torusMap);
-        }
-        else {
-            torusMap = System.getenv("TORUSMAP");
-            if (torusMap!=null) {
-                System.out.println("["+this.getName()+"] Torus Map chosen based on env: "+torusMap);
-            }
-        }
-        if (torusMap==null) {
-            throw new RuntimeException("["+this.getName()+"]  Failed to find torus map name in yaml or env.");
-        }
-        if (solenoidMap!=null) {
-            System.out.println("["+this.getName()+"] solenoid Map chosen based on yaml: "+solenoidMap);
-        }
-        else {
-            solenoidMap = System.getenv("SOLENOIDMAP");
-            if (solenoidMap!=null) {
-                System.out.println("["+this.getName()+"] solenoid Map chosen based on env: "+solenoidMap);
-            }
-        }
-        if (solenoidMap==null) {
-            throw new RuntimeException("["+this.getName()+"]  Failed to find solenoid map name in yaml or env.");
-        }
-        String mapDir = CLASResources.getResourcePath("etc")+"/data/magfield";
-        try {
-            MagneticFields.getInstance().initializeMagneticFields(mapDir,torusMap,solenoidMap);
-        }
-        catch (Exception e) {
-            e.printStackTrace();
-        }
+        super("DCHB");
     }
 
     @Override
     public boolean init() {
-
+        // Load cuts
         Constants.Load();
-        String useSTTConf = this.getEngineConfigString("useStartTime");
-        
-        if (useSTTConf!=null) {
-            System.out.println("["+this.getName()+"] run with start time in tracking config chosen based on yaml ="+useSTTConf);
-            Constants.setUSETSTART(Boolean.valueOf(useSTTConf));
-        }
-        else {
-            useSTTConf = System.getenv("USESTT");
-            if (useSTTConf!=null) {
-                System.out.println("["+this.getName()+"] run with start time in tracking config chosen based on env ="+useSTTConf);
-                Constants.setUSETSTART(Boolean.valueOf(useSTTConf));
-            }
-        }
-        if (useSTTConf==null) {
-             System.out.println("["+this.getName()+"] run with start time in tracking config chosen based on default ="+Constants.isUSETSTART());
-        }
-        
-        this.initializeMagneticFields();
-
-        clasDictionaryPath= CLASResources.getResourcePath("etc");
-        String[]  dcTables = new String[]{
-            "/calibration/dc/signal_generation/doca_resolution",
-          //  "/calibration/dc/time_to_distance/t2d",
-            "/calibration/dc/time_to_distance/time2dist",
-         //   "/calibration/dc/time_corrections/T0_correction",
-            "/calibration/dc/time_corrections/tdctimingcuts",
-            "/calibration/dc/time_jitter",
-            "/calibration/dc/tracking/wire_status",
-        };
-
-        requireConstants(Arrays.asList(dcTables));
-        // Get the constants for the correct variation
-        String dcvarname = CLASResources.getEnvironmentVariable("DCDATABASEVARIATION");
-        String dcvariationName = Optional.ofNullable(dcvarname).orElse("default");
-        this.getConstantsManager().setVariation(dcvariationName);
-
-        // Load the geometry
-        String varname = CLASResources.getEnvironmentVariable("GEOMETRYDATABASEVARIATION");
-        String variationName = Optional.ofNullable(varname).orElse("default");
-
-        ConstantProvider provider = GeometryFactory.getConstants(DetectorType.DC, 11, variationName);
-        dcDetector = new DCGeant4Factory(provider, DCGeant4Factory.MINISTAGGERON);
-        
+        super.initializeMagneticFields();
+        super.setStartTimeOption();
+        super.LoadTables();
         return true;
     }
-    
-    
     @Override
     public boolean processDataEvent(DataEvent event) {
             //setRunConditionsParameters( event) ;
@@ -165,24 +74,15 @@ public class DCHBEngine extends ReconstructionEngine {
             if(timeStamp==-1)
                 return true;
             
-            IndexedTable tabJ=this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_jitter");
+            IndexedTable tabJ=super.getConstantsManager().getConstants(newRun, "/calibration/dc/time_jitter");
             double period = tabJ.getDoubleValue("period", 0,0,0);
             int    phase  = tabJ.getIntValue("phase", 0,0,0);
             int    cycles = tabJ.getIntValue("cycles", 0,0,0);
             
             if(cycles>0) triggerPhase=period*((timeStamp+phase)%cycles); 
-//            if(newRun>1000) {
-//                MagneticFields.getInstance().initializeMagneticFields(clasDictionaryPath+"/data/magfield/", TorusMap.SYMMETRIC);
-//            } else {
-//                MagneticFields.getInstance().initializeMagneticFields(clasDictionaryPath+"/data/magfield/", TorusMap.SYMMETRIC);
-//            }
-            String varname = CLASResources.getEnvironmentVariable("DCDATABASEVARIATION");
-            String variationName = Optional.ofNullable(varname).orElse("default");
+
             TableLoader.FillT0Tables(newRun, variationName);
-            TableLoader.Fill(this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist")); 
-            //CCDBTables.add(this.getConstantsManager().getConstants(newRun, "/calibration/dc/signal_generation/doca_resolution"));
-            //CCDBTables.add(this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/t2d"));
-            //CCDBTables.add(this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_corrections/T0_correction"));
+            TableLoader.Fill(super.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist")); 
             
             double shift =0;
             if(newRun>1890) {
@@ -223,9 +123,9 @@ public class DCHBEngine extends ReconstructionEngine {
 
        HitReader hitRead = new HitReader();
        hitRead.fetch_DCHits(event, noiseAnalysis, parameters, results, 
-               this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), 
-               this.getConstantsManager().getConstants(newRun,"/calibration/dc/time_corrections/tdctimingcuts"), 
-               this.getConstantsManager().getConstants(newRun,"/calibration/dc/tracking/wire_status"), 
+               super.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), 
+               super.getConstantsManager().getConstants(newRun,"/calibration/dc/time_corrections/tdctimingcuts"), 
+               super.getConstantsManager().getConstants(newRun,"/calibration/dc/tracking/wire_status"), 
                dcDetector, triggerPhase);
 
        List<Hit> hits = new ArrayList<Hit>();
@@ -283,7 +183,7 @@ public class DCHBEngine extends ReconstructionEngine {
 
 
         CrossListFinder crossLister = new CrossListFinder();
-        CrossList crosslist = crossLister.candCrossLists(crosses, false, this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), dcDetector, null);
+        CrossList crosslist = crossLister.candCrossLists(crosses, false, super.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), dcDetector, null);
 
         //6) find the list of  track candidates
         TrackCandListFinder trkcandFinder = new TrackCandListFinder("HitBased");
@@ -363,7 +263,7 @@ public class DCHBEngine extends ReconstructionEngine {
         List<Cross> pcrosses = crossMake.find_Crosses(segments, dcDetector);
 
         //
-        CrossList pcrosslist = crossLister.candCrossLists(pcrosses, false, this.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), dcDetector, null);
+        CrossList pcrosslist = crossLister.candCrossLists(pcrosses, false, super.getConstantsManager().getConstants(newRun, "/calibration/dc/time_to_distance/time2dist"), dcDetector, null);
 
         List<Track> mistrkcands =trkcandFinder.getTrackCands(pcrosslist, dcDetector, DCSwimmer.getTorScale());
         if(mistrkcands.size()>0) {
